Equipment: remove some unused code; combine two functions
- acceptsChildren and isContainer are identical. Combine them and pick
  better name.
- remove canAddModifier from EquipmentBuilderFacade, it is unused. We
	should possibly actually use it instead of removing it.

+++ code/src/java/pcgen/core/PlayerCharacter.java
@@ -7635,7 +7635,7 @@ public class PlayerCharacter implements Cloneable, VariableContainer
 
 		if (addAll && mergeItem && (existingSet != null))
 		{
-			newQty = tempQty + getEquippedQty(eSet, eqI).floatValue();
+			newQty = tempQty + getEquippedQty(eSet, eqI);
 			existingSet.setQty(newQty);
 			eqI.setQty(newQty);
 			eqI.setNumberCarried(newQty);
